# raftstore-v2, coprocessor: use snapshot sequence number for cache