--- a/components/engine_panic/src/snapshot.rs
+++ b/components/engine_panic/src/snapshot.rs
@@ -4,6 +4,7 @@ use std::ops::Deref;
 
 use engine_traits::{
     CfNamesExt, IterOptions, Iterable, Iterator, Peekable, ReadOptions, Result, Snapshot,
+    SnapshotMiscExt,
 };
 
 use crate::{db_vector::PanicDbVector, engine::PanicEngine};
@@ -77,6 +78,12 @@ impl Iterator for PanicSnapshotIterator {
     }
 
     fn valid(&self) -> Result<bool> {
+        panic!()
+    }
+}
+
+impl SnapshotMiscExt for PanicSnapshot {
+    fn sequence_number(&self) -> u64 {
         panic!()
     }
 }