--- a/components/engine_rocks/src/snapshot.rs
+++ b/components/engine_rocks/src/snapshot.rs
@@ -7,6 +7,7 @@ use std::{
 
 use engine_traits::{
     self, CfNamesExt, IterOptions, Iterable, Peekable, ReadOptions, Result, Snapshot,
+    SnapshotMiscExt,
 };
 use rocksdb::{rocksdb_options::UnsafeSnap, DBIterator, DB};
 
@@ -101,5 +102,11 @@ impl Peekable for RocksSnapshot {
 impl CfNamesExt for RocksSnapshot {
     fn cf_names(&self) -> Vec<&str> {
         self.db.cf_names()
+    }
+}
+
+impl SnapshotMiscExt for RocksSnapshot {
+    fn sequence_number(&self) -> u64 {
+        unsafe { self.snap.get_sequence_number() }
     }
 }