--- a/components/engine_traits/src/snapshot_misc.rs
+++ b/components/engine_traits/src/snapshot_misc.rs
@@ -1,0 +1,5 @@
+// Copyright 2023 TiKV Project Authors. Licensed under Apache-2.0.
+
+pub trait SnapshotMiscExt {
+    fn sequence_number(&self) -> u64;
+}