--- a/components/raftstore-v2/src/operation/query/local.rs
+++ b/components/raftstore-v2/src/operation/query/local.rs
@@ -211,6 +211,7 @@ where
                             Arc::new(delegate.cached_tablet.cache().snapshot()),
                             region,
                         );
+
                         // Ensures the snapshot is acquired before getting the time
                         atomic::fence(atomic::Ordering::Release);
                         let snapshot_ts = monotonic_raw_now();
@@ -275,6 +276,7 @@ where
                     }
                 };
 
+                snap.set_from_v2();
                 snap.txn_ext = Some(delegate.txn_ext.clone());
                 snap.term = NonZeroU64::new(delegate.term);
                 snap.txn_extra_op = delegate.txn_extra_op.load();