<!--
Thank you for contributing to TiKV! 

If you haven't already, please read TiKV's [CONTRIBUTING](https://github.com/tikv/tikv/blob/master/CONTRIBUTING.md) document. 

If you're unsure about anything, just ask; somebody should be along to answer within a day or two.

PR Title Format:
1. module [, module2, module3]: what's changed
2. *: what's changed
-->

### What is changed and how it works?
<!--

Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and 
linking the relevant issues via the "close" or "ref".

For more info, check https://github.com/tikv/tikv/blob/master/CONTRIBUTING.md#linking-issues.

-->
Issue Number: Ref #14570

What's Changed:

<!--

You could use "commit message" code block to add more description to the final commit message.
For more info, check https://github.com/tikv/tikv/blob/master/CONTRIBUTING.md#format-of-the-commit-message.

-->

We find a read performance regression in internal test, this PR is a quick evaluation for the cop cache impact to read performance in raftstore v2.

```commit-message
In raftstore v2, a snapshot no longer include applied index, so this PR replaces applied index with snapshot sequence number for coprocessor cache.

See https://github.com/tikv/tikv/issues/14570
```

Test results:

### TPCH

This PR
```
Finished
[Summary] Q1: 0.91s
[Summary] Q10: 0.55s
[Summary] Q11: 0.29s
[Summary] Q12: 0.47s
[Summary] Q13: 0.60s
[Summary] Q14: 0.31s
[Summary] Q15: 1.46s
[Summary] Q16: 0.46s
[Summary] Q17: 1.64s
[Summary] Q18: 11.54s
[Summary] Q19: 1.01s
[Summary] Q2: 3.40s
[Summary] Q20: 0.74s
[Summary] Q21: 1.69s
[Summary] Q22: 0.44s
[Summary] Q3: 3.82s
[Summary] Q4: 1.96s
[Summary] Q5: 10.50s
[Summary] Q6: 0.17s
[Summary] Q7: 1.27s
[Summary] Q8: 1.00s
[Summary] Q9: 15.49s
```

Master
```
Finished
[Summary] Q1: 1.52s
[Summary] Q10: 1.15s
[Summary] Q11: 1.90s
[Summary] Q12: 0.96s
[Summary] Q13: 1.95s
[Summary] Q14: 0.87s
[Summary] Q15: 1.49s
[Summary] Q16: 0.93s
[Summary] Q17: 1.69s
[Summary] Q18: 11.23s
[Summary] Q19: 1.42s
[Summary] Q2: 1.14s
[Summary] Q20: 0.95s
[Summary] Q21: 1.75s
[Summary] Q22: 0.76s
[Summary] Q3: 1.02s
[Summary] Q4: 0.77s
[Summary] Q5: 6.88s
[Summary] Q6: 0.63s
[Summary] Q7: 1.51s
[Summary] Q8: 1.41s
[Summary] Q9: 15.77s
```


<!--

```mysql
MySQL [ycsb]> explain analyze select count(*) from usertable;
+----------------------------+-------------+----------+-----------+-----------------+-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+-----------------------------------+-----------+------+
| id                         | estRows     | actRows  | task      | access object   | execution info                                                                                                                                                                                                                                                                                                                                                                | operator info                     | memory    | disk |
+----------------------------+-------------+----------+-----------+-----------------+-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+-----------------------------------+-----------+------+
| HashAgg_11                 | 1.00        | 1        | root      |                 | time:1.61s, loops:2, RU:174102.368426, partial_worker:{wall_time:1.614536734s, concurrency:5, task_num:1, tot_wait:8.071493313s, tot_exec:26.773µs, tot_time:8.071533024s, max:1.614418898s, p95:1.614418898s}, final_worker:{wall_time:0s, concurrency:5, task_num:1, tot_wait:8.072433825s, tot_exec:37.599µs, tot_time:8.072473882s, max:1.614516096s, p95:1.614516096s}   | funcs:count(Column#13)->Column#12 | 12.0 KB   | N/A  |
| └─TableReader_12           | 1.00        | 116      | root      |                 | time:1.61s, loops:2, cop_task: {num: 116, max: 366.4ms, min: 59.2ms, avg: 183.8ms, p95: 310.2ms, max_proc_keys: 181679, p95_proc_keys: 139133, tot_proc: 15.5s, tot_wait: 37.4ms, rpc_num: 116, rpc_time: 21.3s, copr_cache_hit_ratio: 0.00, build_task_duration: 215.1µs, max_distsql_concurrency: 15}                                                                       | data:HashAgg_5                    | 418 Bytes | N/A  |
|   └─HashAgg_5              | 1.00        | 116      | cop[tikv] |                 | tikv_task:{proc max:366ms, min:59ms, avg: 182.7ms, p80:231ms, p95:309ms, iters:9822, tasks:116}, scan_detail: {total_process_keys: 10000287, total_process_keys_size: 11069058766, total_keys: 10000403, get_snapshot_time: 942µs, rocksdb: {key_skipped_count: 20000458, block: {cache_hit_count: 357718, read_count: 3357, read_byte: 75.5 MB, read_time: 176.2ms}}}        | funcs:count(1)->Column#13         | N/A       | N/A  |
|     └─TableFullScan_10     | 10411091.00 | 10000287 | cop[tikv] | table:usertable | tikv_task:{proc max:366ms, min:59ms, avg: 182.4ms, p80:231ms, p95:309ms, iters:9822, tasks:116}                                                                                                                                                                                                                                                                               | keep order:false                  | N/A       | N/A  |
+----------------------------+-------------+----------+-----------+-----------------+-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+-----------------------------------+-----------+------+
4 rows in set (1.617 sec)


MySQL [ycsb]> explain analyze select count(*) from usertable;
+----------------------------+-------------+----------+-----------+-----------------+--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+-----------------------------------+---------+------+
| id                         | estRows     | actRows  | task      | access object   | execution info                                                                                                                                                                                                                                                                                                                                                                 | operator info                     | memory  | disk |
+----------------------------+-------------+----------+-----------+-----------------+--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+-----------------------------------+---------+------+
| HashAgg_11                 | 1.00        | 1        | root      |                 | time:841.9ms, loops:2, RU:102897.414482, partial_worker:{wall_time:841.913015ms, concurrency:5, task_num:1, tot_wait:4.208911203s, tot_exec:27.968µs, tot_time:4.208947095s, max:841.793239ms, p95:841.793239ms}, final_worker:{wall_time:0s, concurrency:5, task_num:1, tot_wait:4.209025048s, tot_exec:15.268µs, tot_time:4.20904279s, max:841.817049ms, p95:841.817049ms}   | funcs:count(Column#13)->Column#12 | 13.5 KB | N/A  |
| └─TableReader_12           | 1.00        | 116      | root      |                 | time:841.7ms, loops:2, cop_task: {num: 116, max: 256.6ms, min: 406.7µs, avg: 85.1ms, p95: 224.9ms, max_proc_keys: 181679, p95_proc_keys: 113238, tot_proc: 8.21s, tot_wait: 16.8ms, rpc_num: 116, rpc_time: 9.87s, copr_cache_hit_ratio: 0.43, build_task_duration: 223.5µs, max_distsql_concurrency: 15}                                                                      | data:HashAgg_5                    | 1.99 KB | N/A  |
|   └─HashAgg_5              | 1.00        | 116      | cop[tikv] |                 | tikv_task:{proc max:362ms, min:56ms, avg: 169ms, p80:218ms, p95:287ms, iters:9822, tasks:116}, scan_detail: {total_process_keys: 5928502, total_process_keys_size: 6562159015, total_keys: 5928568, get_snapshot_time: 1.16ms, rocksdb: {key_skipped_count: 11856938, block: {cache_hit_count: 214021}}}                                                                       | funcs:count(1)->Column#13         | N/A     | N/A  |
|     └─TableFullScan_10     | 10411091.00 | 10000287 | cop[tikv] | table:usertable | tikv_task:{proc max:361ms, min:56ms, avg: 168.7ms, p80:218ms, p95:285ms, iters:9822, tasks:116}                                                                                                                                                                                                                                                                                | keep order:false                  | N/A     | N/A  |
+----------------------------+-------------+----------+-----------+-----------------+--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+-----------------------------------+---------+------+
4 rows in set (0.844 sec)


MySQL [ycsb]> explain analyze select count(*) from usertable;
+----------------------------+-------------+----------+-----------+-----------------+--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+-----------------------------------+---------+------+
| id                         | estRows     | actRows  | task      | access object   | execution info                                                                                                                                                                                                                                                                                                                                                                 | operator info                     | memory  | disk |
+----------------------------+-------------+----------+-----------+-----------------+--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+-----------------------------------+---------+------+
| HashAgg_11                 | 1.00        | 1        | root      |                 | time:844.9ms, loops:2, RU:97648.157242, partial_worker:{wall_time:844.909199ms, concurrency:5, task_num:1, tot_wait:4.222177995s, tot_exec:18.205µs, tot_time:4.222203725s, max:844.444194ms, p95:844.444194ms}, final_worker:{wall_time:0s, concurrency:5, task_num:1, tot_wait:4.222281775s, tot_exec:15.637µs, tot_time:4.222299609s, max:844.465929ms, p95:844.465929ms}   | funcs:count(Column#13)->Column#12 | 13.5 KB | N/A  |
| └─TableReader_12           | 1.00        | 116      | root      |                 | time:844.4ms, loops:2, cop_task: {num: 116, max: 298.9ms, min: 303.4µs, avg: 81.2ms, p95: 239.4ms, max_proc_keys: 181679, p95_proc_keys: 113238, tot_proc: 7.71s, tot_wait: 15.3ms, rpc_num: 116, rpc_time: 9.42s, copr_cache_hit_ratio: 0.47, build_task_duration: 198.9µs, max_distsql_concurrency: 15}                                                                      | data:HashAgg_5                    | 2.26 KB | N/A  |
|   └─HashAgg_5              | 1.00        | 116      | cop[tikv] |                 | tikv_task:{proc max:362ms, min:56ms, avg: 168.8ms, p80:222ms, p95:291ms, iters:9822, tasks:116}, scan_detail: {total_process_keys: 5627669, total_process_keys_size: 6229169980, total_keys: 5627731, get_snapshot_time: 1.05ms, rocksdb: {key_skipped_count: 11255276, block: {cache_hit_count: 203158}}}                                                                     | funcs:count(1)->Column#13         | N/A     | N/A  |
|     └─TableFullScan_10     | 10411091.00 | 10000287 | cop[tikv] | table:usertable | tikv_task:{proc max:361ms, min:56ms, avg: 168.5ms, p80:222ms, p95:291ms, iters:9822, tasks:116}                                                                                                                                                                                                                                                                                | keep order:false                  | N/A     | N/A  |
+----------------------------+-------------+----------+-----------+-----------------+--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+-----------------------------------+---------+------+
4 rows in set (0.847 sec)
```
-->

### Related changes

- Need to cherry-pick to the release branch

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

- Integration test

### Release note <!-- bugfixes or new feature need a release note -->

```release-note
None
```