@tonyxuqqi 2023-06-30
Will a prewrite update the snapshot sequence number?