@overvenus 2023-07-04
> Will a prewrite update the snapshot sequence number?

Yes, it updates both snapshot sequence number and applied index.